+++ src/automatics/calc.py
@@ -2,84 +2,42 @@
 
 from typing import Any
 
-from pydantic import BaseModel, Field, model_validator
-
-from .utils.misc import hash_from_dict
+from pydantic import BaseModel
 
 CalculationDict = dict[str, Any]
 
 
-class Calculation(BaseModel):
-    r"""Calculation input parameters and metadata.
+class Model(BaseModel):
+    r"""Calculation model specification.
 
     Attributes
     ----------
     program : str
         Quantum chemistry program used (psi4, ORCA, ...)
+    program_version : str, optional
+        Quantum chemistry program version.
     calc_type : str
         Calculation type (energy, optimization, ...)
-    method : str, optional
+    method : str
         Computational method (B3LYP, MP2, ...)
     basis : str, optional
-        Basis set.
-    input_data : dict, optional
-        Dictionary containing optional calculation parameters.
-    provenance_source : str, optional
-        Original producer of the provenance data.
-    provenance : dict, optional
-        Dictionary containing optional provenance data.
-    base_hash : str, optional
-        Auto-populated attribute hashing base Calculation attributes.
-    full_hash : str, optional
-        Auto-populated attribute hashing base Calculation attributes and input_data.
+        Orbital basis set.
 
     Example
     -------
     ```
-    calc = Calculation(
+    opt_model = Model(
         program = "orca",
+        program_version = "6.1.1",
         calc_type = "optimization",
         method = "b3lyp",
         basis = "def2-SVP",
-        input_data = {
-            "inp": "%MAXCORE 4000%\nbase 'opt'\n! B3LYP OPT\n\n*xyzfile 0 2 inp.xyz\n"
-        },
-        provenance_source = "custom",
-        provenance = {
-            "program_version": "6.1.1",
-            "aux_basis": "def2/J",
-            "wall_time": "00:00:05:58",
-        }
     )
     ```
     """
 
     program: str
+    program_version: str = ""
     calc_type: str
     method: str
-    basis: str | None = None
-    input_data: dict[str, Any] = Field(default_factory=dict)
-    provenance_source: str | None = None
-    provenance: dict[str, Any] = Field(default_factory=dict)
-
-    base_hash: str | None = None
-    full_hash: str | None = None
-
-    @model_validator(mode="after")
-    def populate_hashes(self) -> "Calculation":
-        """Automatically compute minimal and full hashes on initialization."""
-        base_components = {
-            "program": self.program,
-            "calc_type": self.calc_type,
-            "method": self.method,
-            "basis": self.basis,
-        }
-
-        if self.base_hash is None:
-            self.base_hash = hash_from_dict(base_components)
-
-        if self.full_hash is None:
-            full_components = {**base_components, "input_data": self.input_data}
-            self.full_hash = hash_from_dict(full_components)
-
-        return self
+    basis: str = ""