+++ src/automatics/utils/__init__.py
@@ -1,6 +1,5 @@
 """automatics utilities."""
 
 from . import exceptions, types
-from .misc import hash_from_dict
 
-__all__ = ["exceptions", "hash_from_dict", "types"]
+__all__ = ["exceptions", "types"]